+++ sync-basalam/includes/Actions/Controller/TicketActions/CreateTicket.php
@@ -19,6 +19,10 @@ class CreateTicket extends ActionController
 
         $result = $ticketManager->createTicket($title, $subject, $content);
         if (isset($result['body'])) $ticket = json_decode($result['body'], true);
+        else {
+            wp_die('خطایی در ارسال تیکت رخ داده است. لطفا مجددا تلاش کنید.');
+        }
+        
         wp_redirect(admin_url("admin.php?page=sync_basalam_ticket&ticket_id=" . $ticket['data']['id']));
         exit();
     }